F/Checkout Safe Version

+++ server/db/models/orderDetail.js
@@ -9,7 +9,7 @@ const OrderDetail = db.define('orderDetail', {
       min: 1
     }
   },
-  price: Sequelize.INTEGER
+  price: Sequelize.DECIMAL(10, 2)
 })
 
 OrderDetail.beforeCreate((orderDetail) => {
